+++ tools/slims.py
